requirements.txt
supabase
python-dotenv
icecream
imagekitio
opencv-python
cloudflare
sqlalchemy
alembic
psycopg2-binary
-e .
